# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CornerLab API
  description: >
    Plataforma de inteligência estatística, backtesting e gestão de estratégia
    para o mercado de escanteios. A API nunca recomenda apostas — apenas
    organiza dados históricos, calcula estatísticas e simula cenários.
    O Módulo de Inteligência Estatística (grupo `intelligence`) segue as mesmas
    restrições: nenhum endpoint prevê resultados futuros, promete lucro ou
    recomenda mercados — todo indicador é derivado de dados históricos
    armazenados, com rastreabilidade (liga/temporada/jogos analisados/data de
    atualização) em `meta` nas respostas.
  version: "0.2.0"
servers:
  - url: /api/v1
...
paths:
...
  /intelligence/explain:
    post:
      tags: [intelligence]
      summary: Explicação em linguagem analítica (IA restrita a dados armazenados)
      description: >
        Usa exclusivamente os dados estatísticos já calculados (consistência,
        tendência, estabilidade, score e, se informado, adversário) como
        contexto para o modelo de linguagem. Nunca recomenda apostas, prevê
        resultados ou promete lucro — respostas com linguagem proibida são
        filtradas antes de retornar. Requer ANTHROPIC_API_KEY configurada no
        servidor; retorna 503 se não configurada.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExplainRequest' }
      responses:
        '200': { description: "Resposta em linguagem natural + snapshot dos dados usados" }
        '503': { description: "IA não configurada (ANTHROPIC_API_KEY ausente)" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ExplainRequest:
      type: object
      required: [team_id, league_id]
      properties:
        team_id: { type: integer }
        league_id: { type: integer }
        opponent_id: { type: integer, nullable: true }
        question: { type: string, example: "Por que o Flamengo é considerado consistente na linha de 9.5 escanteios?" }
        limit: { type: integer, default: 10 }
...